Style sidebar and fix sizing issues

+++ Java-Spring-Backend/src/main/java/org/example/summer/dto/UserResultRequest.java
@@ -30,5 +30,5 @@ public class UserResultRequest {
 
     private String topNeighbourhoodName;
 
-    private int topNeighbourhoodRating;
+    private double topNeighbourhoodRating;
 }
